Last type cast conversion fix.

+++ src/shared/hashmap.c
@@ -84,7 +84,7 @@ static void* allocate_tile(struct pool **first_pool, void **first_tile, size_t t
                 n = *first_pool ? (*first_pool)->n_tiles : 0;
                 n = MAX(512U, n * 2);
                 size = PAGE_ALIGN(ALIGN(sizeof(struct pool)) + n*tile_size);
-                n = (size - ALIGN(sizeof(struct pool))) / tile_size;
+                n = (unsigned)((size - ALIGN(sizeof(struct pool))) / tile_size);
 
                 p = malloc0(size);
                 if (!p)
